requirements.txt
# Core
python-dotenv>=1.0
pyyaml>=6.0
pydantic>=2.0
pydantic-settings>=2.0

# ML / Data
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
evidently>=0.4.0
scipy>=1.11.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0

# Monitoring
prometheus-client>=0.17.0

# Storage
redis>=4.5.0
pyarrow>=13.0.0

# Reporting
jinja2>=3.1.0

# Dev
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.21.0
ruff>=0.1
pre-commit>=3.0
